does code cleanup in networkcontroller, equipmentcontroller, refactors networkcontroller
